+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v4.15.1
+🚀Private Release - 4.15.1 🚀
+MercadoPagoSDKV4 - Private Version
+ - Processing Modes Fixes
+
+# v4.15.0
+🚀Private Release - 4.15.0 🚀
+MercadoPagoSDKV4 - Private Version
+ - Gateway Mode Support
+
 # v4.14.0
 🚀Private Release - 4.14.0 🚀
 MercadoPagoSDKV4 - Private Version